Fix: deprecated numpy behavior
Adding an uint64 and an int64 in numpy results in a float64 for whatever reason so cast to uint64

+++ mce.py
@@ -458,7 +458,7 @@ class mce(object):
             btypes += ch.Blocks.ravel()
             counts = bincount(btypes)
 
-            blockCounts[:counts.shape[0]] += counts
+            blockCounts[:counts.shape[0]] += counts.astype(numpy.uint64)
             sizeOnDisk += ch.compressedSize()
             ch.unload()
             if i % 100 == 0:
